requirements.txt
argparse == 1.4.0
requests >= 2.24.0
PySmbClient == 0.1.5
smbprotocol == 1.13.0